backend/pyproject.toml
[tool.poetry]
name = "sagahealth-custodian-api"
version = "0.1.0"
description = "SagaHealth Custodian API"
authors = []
readme = "../README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = ">=0.115.0"
uvicorn = {extras = ["standard"], version = ">=0.30.0"}
httpx = "^0.28.1"
pydantic-settings = "^2.13.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.48"}
bcrypt = "^5.0.0"
aiosqlite = "^0.22.1"
pyotp = ">=2.9.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0.0"
pytest-asyncio = ">=0.24.0"
httpx = ">=0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
